+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/invoice/payment/InvoicePaymentValidateServiceBankPayImpl.java
@@ -145,7 +145,7 @@ public class InvoicePaymentValidateServiceBankPayImpl extends InvoicePaymentVali
     BankOrder bankOrder = bankOrderCreateService.createBankOrder(invoicePayment);
 
     // Disable confirmation of bank order. In the futur we will add an option.
-    //bankOrderService.confirm(bankOrder);
+    // bankOrderService.confirm(bankOrder);
 
     invoicePayment.setBankOrder(bankOrder);